# backend/docs/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TrueAdmin API",
    "version": "0.1.0",
    "description": "Generated from TrueAdmin controller attributes. Code annotations provide default interface metadata."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/admin/organization/roles": {
      "post": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminRoleController_create",
        "summary": "新增角色",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminRoleController@create",
          "permission": "system:role:create",
          "permissionMode": "single",
          "permissions": [
            "system:role:create"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminRoleController_list",
        "summary": "角色列表",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminRoleController@list",
          "permission": "system:role:list",
          "permissionMode": "single",
          "permissions": [
            "system:role:list"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/organization/roles/{id}": {
      "delete": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminRoleController_delete",
        "summary": "删除角色",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminRoleController@delete",
          "permission": "system:role:delete",
          "permissionMode": "single",
          "permissions": [
            "system:role:delete"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminRoleController_detail",
        "summary": "角色详情",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminRoleController@detail",
          "permission": "system:role:detail",
          "permissionMode": "single",
          "permissions": [
            "system:role:detail"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminRoleController_update",
        "summary": "编辑角色",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminRoleController@update",
          "permission": "system:role:update",
          "permissionMode": "single",
          "permissions": [
            "system:role:update"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/organization/roles/{id}/menus": {
      "post": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminRoleController_authorize",
        "summary": "角色菜单授权",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminRoleController@authorize",
          "permission": "system:role:authorize",
          "permissionMode": "single",
          "permissions": [
            "system:role:authorize"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/organization/roles/options": {
      "get": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminRoleController_options",
        "summary": "角色选项",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminRoleController@options",
          "permission": "system:role:list",
          "permissionMode": "single",
          "permissions": [
            "system:role:list"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/organization/roles/tree": {
      "get": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminRoleController_tree",
        "summary": "角色树",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminRoleController@tree",
          "permission": "system:role:list",
          "permissionMode": "single",
          "permissions": [
            "system:role:list"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/organization/users": {
      "post": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminUserController_create",
        "summary": "新增管理员用户",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminUserController@create",
          "permission": "system:user:create",
          "permissionMode": "single",
          "permissions": [
            "system:user:create"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminUserController_list",
        "summary": "管理员用户列表",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminUserController@list",
          "permission": "system:user:list",
          "permissionMode": "single",
          "permissions": [
            "system:user:list"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/organization/users/{id}": {
      "delete": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminUserController_delete",
        "summary": "删除管理员用户",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminUserController@delete",
          "permission": "system:user:delete",
          "permissionMode": "single",
          "permissions": [
            "system:user:delete"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminUserController_detail",
        "summary": "管理员用户详情",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminUserController@detail",
          "permission": "system:user:detail",
          "permissionMode": "single",
          "permissions": [
            "system:user:detail"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminUserController_update",
        "summary": "编辑管理员用户",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminUserController@update",
          "permission": "system:user:update",
          "permissionMode": "single",
          "permissions": [
            "system:user:update"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}